feati10:Improvement in the delete method

+++ src/main/java/com/github/Real_TimeDeliveryTrackingSystem/Real_TimeDeliveryTrackingSystem/service/CustomerService.java
@@ -1,10 +1,12 @@
 package com.github.Real_TimeDeliveryTrackingSystem.Real_TimeDeliveryTrackingSystem.service;
 
+import com.github.Real_TimeDeliveryTrackingSystem.Real_TimeDeliveryTrackingSystem.dto.PasswordDTO;
 import com.github.Real_TimeDeliveryTrackingSystem.Real_TimeDeliveryTrackingSystem.entity.AddressEntity;
 import com.github.Real_TimeDeliveryTrackingSystem.Real_TimeDeliveryTrackingSystem.entity.CustomerEntity;
 import com.github.Real_TimeDeliveryTrackingSystem.Real_TimeDeliveryTrackingSystem.entity.values.AddressVO;
 import com.github.Real_TimeDeliveryTrackingSystem.Real_TimeDeliveryTrackingSystem.exception.AddressNotFoundException;
 import com.github.Real_TimeDeliveryTrackingSystem.Real_TimeDeliveryTrackingSystem.exception.CustomerNotFoundException;
+import com.github.Real_TimeDeliveryTrackingSystem.Real_TimeDeliveryTrackingSystem.exception.InvalidPasswordException;
 import com.github.Real_TimeDeliveryTrackingSystem.Real_TimeDeliveryTrackingSystem.mapper.BuildMapper;
 import com.github.Real_TimeDeliveryTrackingSystem.Real_TimeDeliveryTrackingSystem.repository.CustomerRepository;
 import com.github.Real_TimeDeliveryTrackingSystem.Real_TimeDeliveryTrackingSystem.service.contract.CustomerServiceContract;
@@ -13,7 +15,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class CustomerService implements CustomerServiceContract {
@@ -22,21 +26,33 @@ public class CustomerService implements CustomerServiceContract {
     private static final String ADDRESS_NOT_FOUND_MESSAGE = "Was not found this address associated with this customer," +
             " please verify the fields and try again.";
 
+    private static final String INVALID_PASSWORD_MESSAGE = "The password typed is incorrect," +
+            " please verify and try again.";
+
     private final CustomerRepository customerRepository;
     private final AddressService addressService;
+    private final PasswordEncoder passwordEncoder;
 
     @Autowired
-    public CustomerService(CustomerRepository customerRepository, AddressService addressService) {
+    public CustomerService(CustomerRepository customerRepository, AddressService addressService, PasswordEncoder passwordEncoder) {
         this.customerRepository = customerRepository;
         this.addressService = addressService;
+        this.passwordEncoder = passwordEncoder;
     }
 
 
     @Override
-    public void delete(String email) {
+    @Transactional
+    public void delete(PasswordDTO passwordDTO) {
 
-        CustomerEntity customerEntity = customerRepository.findCustomerByUserEmail(email)
+        CustomerEntity customerEntity = customerRepository.findCustomerByUserEmail(retrieveUserEmail())
                 .orElseThrow(() -> new CustomerNotFoundException(CUSTOMER_NOT_FOUND_MESSAGE));
+
+        if (!passwordEncoder.matches(passwordDTO.getPassword(), customerEntity.getUser().getPassword())) {
+            throw new InvalidPasswordException(INVALID_PASSWORD_MESSAGE);
+
+        }
+
 
         customerRepository.delete(customerEntity);
 
@@ -69,6 +85,10 @@ public class CustomerService implements CustomerServiceContract {
     @Override
     public AddressVO findAddressOfACustomerByItsId(String addressId) {
 
+        CustomerEntity customerEntity = customerRepository.findCustomerByUserEmail(retrieveUserEmail())
+                .orElseThrow(() -> new CustomerNotFoundException(CUSTOMER_NOT_FOUND_MESSAGE));
+
+        verifyIfAddressIdIsAssociatedWithCustomer(addressId, customerEntity);
 
         return addressService.findById(addressId);
     }
